+++ commands/FF_settings.py
@@ -53,7 +53,7 @@ def RunCommand():
     # Update scene
     # =============================================================================
 
-    meshobj: RhinoCableMeshObject = session.scene.find_by_name(name="CableMesh")
+    meshobj: RhinoCableMeshObject = session.find_cablemesh()
     if not meshobj:
         return